requirements.txt
transformers
streamlit
torch
openai
langchain
st-annotated-text
streamlit-drawable-canvas
PyPDF2
python-dotenv
faiss-cpu
streamlit-extras
tiktoken
playsound
elevenlabs
audio-recorder-streamlit
unstructured
requests
scipy
tenacity
tiktoken
termcolor 
pyairtable
